init commit

+++ CarRentN/src/main/java/lk/ijse/carrentn/dto/RentalDiscountDTO.java
@@ -4,18 +4,18 @@ public class RentalDiscountDTO {
     private int rental_discount_id;
     private int rental_id;
     private int discount_id;
-    private int discount_amount_applied;
+    private double discount_amount_applied;
 
     public RentalDiscountDTO() {
     }
 
-    public RentalDiscountDTO(int rental_id, int discount_id, int discount_amount_applied) {
+    public RentalDiscountDTO(int rental_id, int discount_id, double discount_amount_applied) {
         this.rental_id = rental_id;
         this.discount_id = discount_id;
         this.discount_amount_applied = discount_amount_applied;
     }
 
-    public RentalDiscountDTO(int rental_discount_id, int rental_id, int discount_id, int discount_amount_applied) {
+    public RentalDiscountDTO(int rental_discount_id, int rental_id, int discount_id, double discount_amount_applied) {
         this.rental_discount_id = rental_discount_id;
         this.rental_id = rental_id;
         this.discount_id = discount_id;
@@ -46,11 +46,11 @@ public class RentalDiscountDTO {
         this.discount_id = discount_id;
     }
 
-    public int getDiscount_amount_applied() {
+    public double getDiscount_amount_applied() {
         return discount_amount_applied;
     }
 
-    public void setDiscount_amount_applied(int discount_amount_applied) {
+    public void setDiscount_amount_applied(double discount_amount_applied) {
         this.discount_amount_applied = discount_amount_applied;
     }
